serve/requirements.txt
openai>=1.30.0
pillow
pdf2image
python-dotenv
